daemon: Move containerd inside images
