+++ daemon/images/containerd/image_delete.go
