+++ daemon/images/containerd/image_delete_test.go
