+++ daemon/images/containerd/image_exporter.go
