+++ daemon/images/containerd/image_inspect.go
