+++ daemon/images/containerd/image_manifest.go
