+++ daemon/images/containerd/image_prune.go
