+++ daemon/images/containerd/image_snapshot.go
