+++ daemon/images/containerd/image_snapshot_windows.go
