+++ daemon/images/containerd/image_squash.go
