+++ daemon/images/containerd/imagespec.go
