+++ daemon/images/containerd/resolver.go
