+++ daemon/images/containerd/soft_delete.go
